Adds queryform option to search for records modified by a user, show last modified by
This provides two new features to the occurrence table view and associated search
* A column to show the user who last modified each record (if any), and the ability to sort by this column
* A way to search for records modified by a user that extends the existing interface for showing records entered by a user. Note that this includes any record modified by a user, whether or not they were the last user to modify it. So, in some cases, the last modified by column (above) will show a different user.

Both of these are useful for monitoring the work of student transcribers or other personnel with database edit access. It is intended to be complementary with the Review/Verify Occurrence Edits tool, and the User Stats reports tool, but unique to this implementation is the ability to do more fine-grained searches.