Adds glossary tooltips to taxon profile page descriptions
Adds a new glossary API file at glossary/getterms.php. When called with no arguments, it returns an array of glossary terms. When called with a term ID, it returns the relevant data for that term.

The taxa React page gets the list of glossary terms from the API. It then does a regex search for all of those terms against the description text, and replaces any words that match with html that enables the tooltips. The tooltips are built using JQuery UI tooltips built in taxa/index.php.