Taxon thumbnails now display again in checklists
Arthur turned this off, or changed things by disabling populateImages() in the TaxaManager class construct. However, that code was extremely inefficient, pulling all the images and their data for a taxon (with multiple table joins) just to get the thumbnail URL of the first image in sort sequence. I rewrote the thumbnail code, but it's still inefficient, running a separate query for each taxon.