[classes][garden] move thumbnail URLs into a separate query to fix memory issues
IdentManager is causing slow page loads for pages that display taxon thumbnails (e.g. garden). This is because the imgid to display is selected in a multi-table join along with taxavernaculars and taxstatus, which causes hundreds of thousands of rows to be loaded from the db and eats up huge amounts of memory. Intermittently the page fails to load entirely because the PHP process runs out of memory.

To fix this, we instead do a separate query for just the thumbnail URL to display for each taxon. We do a separate query for each taxon to take advantage of LIMIT 1 in the SQL query, since we only need the first matching row. This avoids, for some taxa, selecting thousands of unused rows from the image table.

The downside of this approach is that it queries the database once per taxon, rather than once overall, in order to get all the thumbnails. However, in the current code, we are actually still querying the database once per taxon AFTER the giant multi-table join, as this query gives us only the imgid. (This extra per-taxon query is on the deleted line 488 of garden/rpc/api.php -- sneaky.)

Since the PHP code no longer needs to keep 300,000+ rows in memory, we no longer need the amped-up memory limit and infinite timeout we are currently using for this query.