[webpack] make all entrypoints depend on header to reduce duplicated chunks (dependencies) in built files
Initially I added a separate entrypoint called 'shared' which would need to be loaded alongside each entrypoint, but since each page loads header already, it's way simpler and more efficient to just have all the other entrypoints depend on header. This means that any chunks included in header (such as react, react-dom, etc.) are now ONLY in header and not duplicated in the other entrypoints.