[profile] when logging out, also delete cookies with path set to /profile
After upgrading, I experienced a weird issue where a PHPSESSID cookie was already set in my browser with a path of /profile, and it wasn't cleared upon logging out, so I ended up in a weird in-between state where I was logged out but the header still acted as if I were logged in. I couldn't see an easy way out of this state as an end user. I'm not sure exactly how this cookie was set initially, perhaps the old Symbiota code used a cookie path sometimes.

Adding these lines and logging out again fixed the issue for me. I think we should leave it for a little while -- not indefinitely, but long enough for everyone who was logged into a session pre-Symbiota 3.1 can log out again and have their cookies reset.