[map] make LeafletSingleClusterMapGroup closer to original by keeping individual layer groups and just having a single group-wide cluster layer
This improves performance without having to change the usage of this class at all. In the previous implementation, the cluster layer was regenerated each time removeLayer and addLayer were called in sequence. In fixing this, I've opted to make the implementation closer to the original rather than diverging further from symbiota.